Fix underestimated saliency increase; add plots.
The original submission had an off-by-one error due to which
the reported final max saliency was not from the final step
of search, but from the penultimate one. Thus, saliency was
in reality increased by more than what was reported in the
document. This commit fixes the reported numbers and adds saliency
plots.